- status icon works too - removed some warnings in listctrls

+++ src/channel/channellistctrl.cpp
@@ -158,7 +158,7 @@ wxString ChannelListctrl::OnGetItemText(long item, long column) const
 void ChannelListctrl::SetTipWindowText(const long item_hit, const wxPoint position)
 {
     int coloumn = getColoumnFromPosition(position);
-    if (coloumn > (int)m_colinfovec.size() || coloumn < 0 || item_hit < 0 || item_hit > m_data.size() )
+    if (coloumn > (int)m_colinfovec.size() || coloumn < 0 || item_hit < 0 || item_hit > (long)m_data.size() )
     {
         m_tiptext = _T("");
     }
